Refactor pipeline file structure, make modes per-pipeline
