+++ orchestration/dagster_orchestration/hca_orchestration/pipelines/__init__.py
@@ -1,0 +1,8 @@
+from hca_orchestration.pipelines.stage_data import stage_data
+from hca_orchestration.pipelines.validate_egress import validate_egress
+
+
+__all__ = [
+    stage_data,
+    validate_egress,
+]
